reader: match against a record using the same field names as the part
Match by looking at the path of the first part; this assumes that the hash
passed into match? does not contain fields that vary amongst the parts of a
rectype.

This generalizes matching and makes it easier to match against fields with
odd names

As a side effect, rename :satz to :sid in all the model classes for
matching

   * lib/gdv/format/reader.rb (match?): use the first part's path for
     matching; (cond_match): tolerate nil's
   * lib/gdv/model/contract.rb, lib/gdv/model/grammar.rb,
     lib/gdv/model/package.rb, lib/gdv/model/partner.rb,
     lib/gdv/model/sparte/kfz.rb, lib/gdv/model/transmission.rb: rename
     :satz to :sid
   * tests/tc_reader.rb (test_match): rename :satz to :sid, add a couple
     more tests